Combined common lexer fixtures
