+++ server/@antlr4-tsplang/test/tspLexer.test.ts
@@ -13,38 +13,14 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-import { ANTLRInputStream, CommonToken } from "antlr4ts"
+import { ANTLRInputStream } from "antlr4ts"
 import { expect } from "chai"
 import "mocha"
 
 import { TspLexer } from "../out/TspLexer.generated"
 
 import { ERROR_THROWER } from "./errorListener.fixture"
-
-// An infinitely nested string array type.
-type NestedArray<T> = Array<string | T>
-/* eslint-disable-next-line @typescript-eslint/no-empty-interface */
-interface RecursiveStringArray extends NestedArray<RecursiveStringArray> {}
-
-// Remove all dimensionality from the recursive array. N-D array -> 1-D array.
-function flatten(value: RecursiveStringArray): Array<string> {
-    const result = new Array<string>()
-    value.forEach(v => {
-        if (typeof v === "object") {
-            result.push(...flatten(v))
-        } else {
-            result.push(v)
-        }
-    })
-    return result
-}
-
-// Array<CommonToken>.map helper function to extract the token type.
-const typeExtractor = (token: CommonToken): number => token.type
-// Array<TokenType>.map helper function to extract the vocab name.
-const nameExtractor = function(ttype: number): string {
-    return this.vocabulary.getSymbolicName(ttype)
-}
+import { flatten, nameExtractor, RecursiveStringArray, typeExtractor } from "./lexer.fixture"
 
 // Test Template
 interface LexTest {
